test(server): share the guard GraphQL harness between the two specs
TestResolver, FeatureModule, RootModule and runGuardedQuery were duplicated
verbatim across user-auth.guard.spec and workspace-auth.guard.spec, so a fix to
the harness had to be made twice.

The resolver's guard is now the parameter, which means the module classes are
built per call rather than at module scope. Each call already stood up its own
Nest app, so the isolation is unchanged.
